+++ lib/chat_web/live/main_live/page/admin_panel_router.ex
@@ -69,8 +69,8 @@ defmodule ChatWeb.MainLive.Page.AdminPanelRouter do
       {:weight_sensor_connection, status} ->
         socket |> AdminPanel.weight_sensor_connection_status(status)
 
-      {:create_cargo_user, name} ->
-        socket |> AdminPanel.create_cargo_user(name)
+      {:create_cargo_user, data} ->
+        socket |> AdminPanel.create_cargo_user(data)
     end
   end