+++ CHANGELOG.md
@@ -15,6 +15,7 @@ straightforward as possible.
 - Router support
 - Upload by Drag&Drop to chats/rooms
 - chat links
+- Import and backup cargo user keys
 
 ### Changed
 - mirroring disabled